Change datatype - get pledge number from UI

+++ src/main/java/com/dkthanh/demo/domain/OptionEntity.java
@@ -19,9 +19,9 @@ public class OptionEntity {
     @Column(name = "option_description", nullable = true, length = 255)
     private String optionDescription;
     @Column(name = "fund_max", nullable = true, precision = 0)
-    private Integer fundMax;
+    private Long fundMax;
     @Column(name = "fund_min", nullable = true, precision = 0)
-    private Integer fundMin;
+    private Long fundMin;
 
     @ManyToOne
     @JoinColumn(name = "project_id")
